+++ src/Components/Help.jsx
@@ -1,0 +1,43 @@
+function Contact() {
+  return (
+    <div>
+      <h1>Help and Support</h1>
+      <div className="divider" />
+      <section className="flow help">
+        <h2>FAQs</h2>
+        <div>
+          <h3>What is your Customer Care Number?</h3>
+          <p>
+            Our customer care number is 00100200210. you acn email us at
+            rasmeetkour735@gmail.com
+          </p>
+        </div>
+        <div>
+          <h3>Did not receive OTP?</h3>
+          <p>
+            Please check if your app is due for an update. If not, please share
+            the details via support@raskour.in
+          </p>
+        </div>
+        <div>
+          <h3>Is there a minimum order value?</h3>
+          <p>
+            We have no minimum order value and you can order for any amount.{" "}
+          </p>
+        </div>
+        <div>
+          <h3>I want to cancel my order</h3>
+          <p>
+            I want to cancel my order We will do our best to accommodate your
+            request if the order is not placed to the restaurant (Customer
+            service number: 080-67466729). Please note that we will have a right
+            to charge a cancellation fee up to full order value to compensate
+            our restaurant and delivery partners if your order has been
+            confirmed.
+          </p>
+        </div>
+      </section>
+    </div>
+  );
+}
+export default Contact;
